requirements.txt
selenium
djangorestframework